cli: move chat_completions tests to the unit/ dir
Following the repo's #[path] unit convention (e.g. src/trie.rs ->
unit/trie_test.rs), lift the test module out of chat_completions.rs into
crates/cli/unit/server/chat_completions_test.rs, linked via
#[cfg(test)] #[path] mod tests;. cli is a binary crate with no lib, so the
module stays in-crate to keep access to the private request helpers;
same 9 tests, no behavior change.
