+++ crates/cli/src/server/chat_completions.rs
@@ -526,159 +526,5 @@ pub async fn handle_chat_completions(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn request(json: &str) -> ChatCompletionRequest {
-        serde_json::from_str(json).expect("valid request json")
-    }
-
-    fn reply_config(json: &str) -> ChatReplyConfig {
-        build_reply_config(&request(json)).expect("valid reply config")
-    }
-
-    #[cfg(not(feature = "capability-grammar"))]
-    fn reply_config_error(json: &str) -> ResponseFormatError {
-        build_reply_config(&request(json)).expect_err("invalid reply config")
-    }
-
-    #[test]
-    fn response_format_maps_to_grammar() {
-        assert!(reply_config(r#"{"messages":[]}"#).grammar.is_none());
-        assert!(reply_config(r#"{"messages":[],"response_format":{"type":"text"}}"#).grammar.is_none());
-
-        #[cfg(feature = "capability-grammar")]
-        assert_eq!(
-            reply_config(r#"{"messages":[],"response_format":{"type":"json_object"}}"#).grammar,
-            Some(Grammar::JsonAny {})
-        );
-    }
-
-    #[test]
-    fn response_format_rejects_grammar_without_capability() {
-        #[cfg(not(feature = "capability-grammar"))]
-        {
-            assert_eq!(
-                reply_config_error(r#"{"messages":[],"response_format":{"type":"json_object"}}"#),
-                ResponseFormatError::GrammarUnsupported
-            );
-            assert_eq!(
-                reply_config_error(
-                    r#"{"messages":[],"response_format":{"type":"json_schema","json_schema":{"schema":{"type":"object"}}}}"#
-                ),
-                ResponseFormatError::GrammarUnsupported
-            );
-        }
-    }
-
-    #[test]
-    fn response_format_unrecognized_is_invalid() {
-        let error = build_reply_config(&request(r#"{"messages":[],"response_format":{"type":"totally-bogus"}}"#))
-            .expect_err("unrecognized response_format should be rejected");
-        assert!(
-            matches!(error, ResponseFormatError::InvalidResponseFormat(_)),
-            "expected InvalidResponseFormat, got {error:?}"
-        );
-    }
-
-    #[cfg(feature = "capability-grammar")]
-    #[test]
-    fn response_format_json_schema_maps_to_grammar() {
-        let config = reply_config(
-            r#"{"messages":[],"response_format":{"type":"json_schema","json_schema":{"name":"person","schema":{"type":"object"}}}}"#,
-        );
-        assert_eq!(
-            config.grammar,
-            Some(Grammar::JsonSchema {
-                schema: r#"{"type":"object"}"#.to_string(),
-            })
-        );
-    }
-
-    #[test]
-    fn response_format_json_schema_rejects_invalid_schema() {
-        let error = build_reply_config(&request(
-            r#"{"messages":[],"response_format":{"type":"json_schema","json_schema":{"schema":{"type":"not-a-json-schema-type"}}}}"#,
-        ))
-        .expect_err("an invalid JSON Schema should be rejected");
-        assert!(
-            matches!(error, ResponseFormatError::InvalidJsonSchema(_)),
-            "expected InvalidJsonSchema, got {error:?}"
-        );
-        assert_eq!(error.code(), "invalid_json_schema");
-    }
-
-    #[test]
-    fn response_format_validation_errors_are_request_errors() {
-        match request_error_response(ResponseFormatError::GrammarUnsupported) {
-            ChatCompletionResult::Error(_) => {},
-            ChatCompletionResult::Json(_) | ChatCompletionResult::Stream(_) => {
-                panic!("response_format validation errors should be request errors")
-            },
-        }
-    }
-
-    #[test]
-    fn malformed_response_format_passes_json_extraction() {
-        for body in [
-            r#"{"messages":[],"response_format":{"type":"totally-bogus"}}"#,
-            r#"{"messages":[],"response_format":"not-even-an-object"}"#,
-        ] {
-            serde_json::from_str::<ChatCompletionRequest>(body)
-                .unwrap_or_else(|error| panic!("expected {body} to pass extraction, got {error}"));
-        }
-    }
-
-    #[rocket::get("/err")]
-    fn err_route() -> ChatCompletionResult {
-        request_error_response(ResponseFormatError::InvalidResponseFormat("bad".to_string()))
-    }
-
-    #[test]
-    fn error_responder_yields_http_400_with_openai_body() {
-        let client = rocket::local::blocking::Client::tracked(rocket::build().mount("/", rocket::routes![err_route]))
-            .expect("rocket client");
-        let response = client.get("/err").dispatch();
-
-        assert_eq!(response.status(), Status::BadRequest);
-        let body: serde_json::Value = response.into_json().expect("json error body");
-        assert_eq!(body["error"]["type"], "invalid_request_error");
-        assert_eq!(body["error"]["param"], "response_format");
-        assert_eq!(body["error"]["code"], "invalid_response_format");
-        assert!(
-            body["error"]["message"].as_str().is_some_and(|message| !message.is_empty()),
-            "expected a non-empty error message, got {body}"
-        );
-    }
-
-    #[cfg(feature = "capability-grammar")]
-    #[test]
-    fn response_format_composes_with_sampling_options() {
-        let stochastic = reply_config(
-            r#"{"messages":[],"temperature":0.7,"top_p":0.9,"top_k":40,"response_format":{"type":"json_object"}}"#,
-        );
-        assert_eq!(stochastic.grammar, Some(Grammar::JsonAny {}));
-        assert_eq!(
-            stochastic.sampling_policy,
-            uzu::types::basic::SamplingPolicy::Custom {
-                method: SamplingMethod::Stochastic {
-                    temperature: Some(0.7),
-                    top_k: Some(40),
-                    top_p: Some(0.9),
-                    min_p: None,
-                    repetition_penalty: None,
-                    suffix_repetition_length: None,
-                },
-            }
-        );
-
-        let greedy = reply_config(r#"{"messages":[],"temperature":0,"response_format":{"type":"json_object"}}"#);
-        assert_eq!(greedy.grammar, Some(Grammar::JsonAny {}));
-        assert_eq!(
-            greedy.sampling_policy,
-            uzu::types::basic::SamplingPolicy::Custom {
-                method: SamplingMethod::Greedy {},
-            }
-        );
-    }
-}
+#[path = "../../unit/server/chat_completions_test.rs"]
+mod tests;
